refactor : reservationType 정리
